Bug # ????? : iOS and Server Mobile "Unbatching"
We're still batching beacons together for a single transmit, but instead of sending summarized records of similar data, we're now sending all of the raw data points.  This allows us to log all "beacons" in the batch as individual beacons in raw S3 log files.

The iOS Beacon classes were reorganized and standardized to the new Beacon Types.  Each Beacon type is a sub-class of the MPBeacon class, and adds its own methods for serialization when called to.

All of the pre-summariziation code (histogram generation, etc) has been removed, since we're now only sending raw data points.

Batches are now gzipped for transport.

The code is now using ObjClean for standardization, though this isn't enforced in the build (yet).  I didn't clean up all of the files -- only files that I touched in a significant way.

A/B test support was added to mPulse as well.

All of the tests were reworked.  The networking request no longer wait for a fixed timeout, but they poll waiting for a beacon with a max timeout.  Tests were added for each new beacon type as well.

Reviewer: Mukul, Andreas, Matt
Document [y/n]: n